pyproject.toml
[tool.poetry]
name = "sam2terra"
version = "0.1.0"
description = ""
authors = ["Saurav Panda <sgp65@cornell.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
pyyaml = "^6.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
